requirements.txt
numpy>=1.21.0
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
requests>=2.25.0
cryptography>=3.4.0
psutil>=5.8.0
redis>=4.0.0
msgpack>=1.0.0
lz4>=3.1.0
xxhash>=2.0.0
multiprocessing-logging>=0.3.0
pyyaml>=6.0
aiohttp>=3.8.0
websockets>=10.0
transformers>=4.20.0
tokenizers>=0.12.0
datasets>=2.0.0
accelerate>=0.20.0